+++ src/main/java/com/example/finalproj/ml/BookML/BookMLEventListener.java
@@ -1,4 +1,4 @@
-package com.example.finalproj.ml.service;
+package com.example.finalproj.ml.BookML;
 
 import com.example.finalproj.Book.service.BookService;
 import org.springframework.context.event.EventListener;
