+++ src/main/java/com/example/finalproj/ml/BookML/BookMLProcessingCompletedEvent.java
@@ -1,4 +1,4 @@
-package com.example.finalproj.ml.service;
+package com.example.finalproj.ml.BookML;
 
 import org.springframework.context.ApplicationEvent;
